requirements.txt
allure-pytest==2.14.1
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
